initial commit to nft_card branch
